Closes issue #61, the final big file is split off into a package.
