+++ xga/sas/__init__.py
@@ -1,0 +1,8 @@
+#  This code is a part of XMM: Generate and Analyse (XGA), a module designed for the XMM Cluster Survey (XCS).
+#  Last modified by David J Turner (david.turner@sussex.ac.uk) 01/09/2020, 17:08. Copyright (c) David J Turner
+
+from .phot import *
+from .spec import *
+
+
+
